migrating to using JDK-22ea, simplify code structure and make improvements

+++ common/src/main/java/cn/zorcc/common/network/MuxEvent.java
@@ -5,6 +5,6 @@ package cn.zorcc.common.network;
  */
 public record MuxEvent(
         int socket,
-        long event
+        int event
 ) {
 }
